+++ example/lib/chat/common/api/api_provider.dart
@@ -243,6 +243,10 @@ class ApiProvider {
     //print(env['API_KEY']);
     return env['API_KEY'];
   }
+  static Future<String> getEnvironmentKeyByName({required String keyName}) async {
+    final env = await parseStringToMap(assetsFileName: '.env');
+    return env[keyName]?? "";
+  }
 
   static Future<Map<String, String>> parseStringToMap(
       {String assetsFileName = '.env'}) async {