+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.2
+* Added support for iOS platform
+* Listen to newly added message in a conversation and token expiration
+* Added method to initialize twilio client
+* Updated documentation
+* 
 ## 1.0.1
 * Bug fixes
 * Added method to get list of participants in a conversation